requirements.txt
numpy==1.19.5
Pillow==8.0.1
opencv-python==4.5.2.54
dlib==19.22.0
face-recognition==1.3.0